Move block worker to common Chain struct
