+++ code/go/0chain.net/sharder/worker.go
@@ -2,10 +2,8 @@ package sharder
 
 import (
 	"context"
-	"sync"
 	"time"
 
-	"0chain.net/core/config"
 	"go.uber.org/zap"
 
 	"0chain.net/chaincore/block"
@@ -46,235 +44,235 @@ func SetupWorkers(ctx context.Context) {
 }
 
 /*BlockWorker - stores the blocks */
-func (sc *Chain) BlockWorker(ctx context.Context) {
-	const stuckDuration = 3 * time.Second
-	var (
-		endRound   int64
-		syncing    bool
-		syncTimer  time.Time
-		timingSync bool
+// func (sc *Chain) BlockWorker(ctx context.Context) {
+// 	const stuckDuration = 3 * time.Second
+// 	var (
+// 		endRound   int64
+// 		syncing    bool
+// 		syncTimer  time.Time
+// 		timingSync bool
 
-		syncBlocksTimer  = time.NewTimer(7 * time.Second)
-		aheadN           = int64(config.GetLFBTicketAhead())
-		maxRequestBlocks = aheadN
+// 		syncBlocksTimer  = time.NewTimer(7 * time.Second)
+// 		aheadN           = int64(config.GetLFBTicketAhead())
+// 		maxRequestBlocks = aheadN
 
-		// triggered after sync process is started
-		stuckCheckTimer = time.NewTimer(10 * time.Second)
-	)
+// 		// triggered after sync process is started
+// 		stuckCheckTimer = time.NewTimer(10 * time.Second)
+// 	)
 
-	for {
-		select {
-		case <-ctx.Done():
-			logging.Logger.Error("BlockWorker exit", zap.Error(ctx.Err()))
-			return
-		case <-stuckCheckTimer.C:
-			logging.Logger.Debug("process block, detected stuck, trigger sync",
-				zap.Int64("round", sc.GetCurrentRound()),
-				zap.Int64("lfb", sc.GetLatestFinalizedBlock().Round))
-			stuckCheckTimer.Reset(stuckDuration)
-			// trigger sync
-			syncBlocksTimer.Reset(0)
-		case <-syncBlocksTimer.C:
-			// reset sync timer to 1 minute
-			syncBlocksTimer.Reset(time.Minute)
+// 	for {
+// 		select {
+// 		case <-ctx.Done():
+// 			logging.Logger.Error("BlockWorker exit", zap.Error(ctx.Err()))
+// 			return
+// 		case <-stuckCheckTimer.C:
+// 			logging.Logger.Debug("process block, detected stuck, trigger sync",
+// 				zap.Int64("round", sc.GetCurrentRound()),
+// 				zap.Int64("lfb", sc.GetLatestFinalizedBlock().Round))
+// 			stuckCheckTimer.Reset(stuckDuration)
+// 			// trigger sync
+// 			syncBlocksTimer.Reset(0)
+// 		case <-syncBlocksTimer.C:
+// 			// reset sync timer to 1 minute
+// 			syncBlocksTimer.Reset(time.Minute)
 
-			var (
-				lfbTk = sc.GetLatestLFBTicket(ctx)
-				lfb   = sc.GetLatestFinalizedBlock()
-			)
+// 			var (
+// 				lfbTk = sc.GetLatestLFBTicket(ctx)
+// 				lfb   = sc.GetLatestFinalizedBlock()
+// 			)
 
-			cr := sc.GetCurrentRound()
-			if cr < lfb.Round {
-				sc.SetCurrentRound(lfb.Round)
-				cr = lfb.Round
-			}
+// 			cr := sc.GetCurrentRound()
+// 			if cr < lfb.Round {
+// 				sc.SetCurrentRound(lfb.Round)
+// 				cr = lfb.Round
+// 			}
 
-			if lfb.Round+aheadN <= cr {
-				logging.Logger.Debug("process block, synced to lfb+ahead, start to force finalize rounds",
-					zap.Int64("current round", cr),
-					zap.Int64("lfb", lfb.Round),
-					zap.Int64("lfb+ahead", lfb.Round+aheadN))
+// 			if lfb.Round+aheadN <= cr {
+// 				logging.Logger.Debug("process block, synced to lfb+ahead, start to force finalize rounds",
+// 					zap.Int64("current round", cr),
+// 					zap.Int64("lfb", lfb.Round),
+// 					zap.Int64("lfb+ahead", lfb.Round+aheadN))
 
-				for rn := lfb.Round + 1; rn <= cr; rn++ {
-					if r := sc.GetRound(rn); r != nil {
-						sc.FinalizeRound(sc.GetRound(rn))
-					}
-				}
-				continue
-			}
+// 				for rn := lfb.Round + 1; rn <= cr; rn++ {
+// 					if r := sc.GetRound(rn); r != nil {
+// 						sc.FinalizeRound(sc.GetRound(rn))
+// 					}
+// 				}
+// 				continue
+// 			}
 
-			endRound = lfbTk.Round + aheadN
+// 			endRound = lfbTk.Round + aheadN
 
-			if endRound <= cr || lfb.Round >= lfbTk.Round {
-				if timingSync {
-					syncCatchupTime.Update(time.Since(syncTimer).Microseconds())
-					timingSync = false
-				}
+// 			if endRound <= cr || lfb.Round >= lfbTk.Round {
+// 				if timingSync {
+// 					syncCatchupTime.Update(time.Since(syncTimer).Microseconds())
+// 					timingSync = false
+// 				}
 
-				logging.Logger.Debug("process block, synced already, continue...")
-				continue
-			}
+// 				logging.Logger.Debug("process block, synced already, continue...")
+// 				continue
+// 			}
 
-			logging.Logger.Debug("process block, sync triggered",
-				zap.Int64("lfb", lfb.Round),
-				zap.Int64("lfb ticket", lfbTk.Round),
-				zap.Int64("current round", cr),
-				zap.Int64("end round", endRound))
+// 			logging.Logger.Debug("process block, sync triggered",
+// 				zap.Int64("lfb", lfb.Round),
+// 				zap.Int64("lfb ticket", lfbTk.Round),
+// 				zap.Int64("current round", cr),
+// 				zap.Int64("end round", endRound))
 
-			// trunc to send maxRequestBlocks each time
-			reqNum := endRound - cr
-			if reqNum > maxRequestBlocks {
-				reqNum = maxRequestBlocks
-			}
+// 			// trunc to send maxRequestBlocks each time
+// 			reqNum := endRound - cr
+// 			if reqNum > maxRequestBlocks {
+// 				reqNum = maxRequestBlocks
+// 			}
 
-			endRound = cr + reqNum
-			syncing = true
-			if !timingSync {
-				timingSync = true
-				syncTimer = time.Now()
-			}
+// 			endRound = cr + reqNum
+// 			syncing = true
+// 			if !timingSync {
+// 				timingSync = true
+// 				syncTimer = time.Now()
+// 			}
 
-			logging.Logger.Debug("process block, sync blocks",
-				zap.Int64("start round", cr+1),
-				zap.Int64("end round", cr+reqNum+1))
-			go sc.requestBlocks(ctx, cr, reqNum)
-		default:
-			cr := sc.GetCurrentRound()
-			lfb := sc.GetLatestFinalizedBlock()
-			bItem, ok := sc.blockBuffer.First()
-			if !ok {
-				// no block in buffer to process
-				time.Sleep(100 * time.Millisecond)
-				continue
-			}
+// 			logging.Logger.Debug("process block, sync blocks",
+// 				zap.Int64("start round", cr+1),
+// 				zap.Int64("end round", cr+reqNum+1))
+// 			go sc.requestBlocks(ctx, cr, reqNum)
+// 		default:
+// 			cr := sc.GetCurrentRound()
+// 			lfb := sc.GetLatestFinalizedBlock()
+// 			bItem, ok := sc.blockBuffer.First()
+// 			if !ok {
+// 				// no block in buffer to process
+// 				time.Sleep(100 * time.Millisecond)
+// 				continue
+// 			}
 
-			b := bItem.Data.(*block.Block)
+// 			b := bItem.Data.(*block.Block)
 
-			logging.Logger.Debug("process block, received block",
-				zap.Int64("block round", b.Round))
-			stuckCheckTimer.Reset(stuckDuration)
-			if b.Round > lfb.Round+aheadN {
-				// trigger sync process to pull the latest blocks when
-				// current round is > lfb.Round + aheadN to break the stuck if any.
-				if !syncing {
-					syncBlocksTimer.Reset(0)
-				}
+// 			logging.Logger.Debug("process block, received block",
+// 				zap.Int64("block round", b.Round))
+// 			stuckCheckTimer.Reset(stuckDuration)
+// 			if b.Round > lfb.Round+aheadN {
+// 				// trigger sync process to pull the latest blocks when
+// 				// current round is > lfb.Round + aheadN to break the stuck if any.
+// 				if !syncing {
+// 					syncBlocksTimer.Reset(0)
+// 				}
 
-				// avoid the skipping logs when syncing blocks
-				if b.Round <= lfb.Round+2*aheadN {
-					logging.Logger.Debug("process block skip",
-						zap.Int64("block round", b.Round),
-						zap.Int64("current round", cr),
-						zap.Int64("lfb", lfb.Round),
-						zap.Bool("syncing", syncing))
-				}
+// 				// avoid the skipping logs when syncing blocks
+// 				if b.Round <= lfb.Round+2*aheadN {
+// 					logging.Logger.Debug("process block skip",
+// 						zap.Int64("block round", b.Round),
+// 						zap.Int64("current round", cr),
+// 						zap.Int64("lfb", lfb.Round),
+// 						zap.Bool("syncing", syncing))
+// 				}
 
-				continue
-			}
+// 				continue
+// 			}
 
-			sc.blockBuffer.Pop()
-			if err := sc.processBlock(ctx, b); err != nil {
-				logging.Logger.Error("process block failed",
-					zap.Error(err),
-					zap.Int64("round", b.Round),
-					zap.String("block", b.Hash),
-					zap.String("prev block", b.PrevHash))
+// 			sc.blockBuffer.Pop()
+// 			if err := sc.processBlock(ctx, b); err != nil {
+// 				logging.Logger.Error("process block failed",
+// 					zap.Error(err),
+// 					zap.Int64("round", b.Round),
+// 					zap.String("block", b.Hash),
+// 					zap.String("prev block", b.PrevHash))
 
-				var pb *block.Block
-				if err == ErrNoPreviousBlock {
-					// fetch the previous block
-					pb, _ = sc.GetNotarizedBlock(ctx, b.PrevHash, b.Round-1)
-				} else if ErrNoPreviousState.Is(err) {
-					// get the previous block from local
-					pb, _ = sc.GetBlock(ctx, b.PrevHash)
-				} else {
-					continue
-				}
+// 				var pb *block.Block
+// 				if err == ErrNoPreviousBlock {
+// 					// fetch the previous block
+// 					pb, _ = sc.GetNotarizedBlock(ctx, b.PrevHash, b.Round-1)
+// 				} else if ErrNoPreviousState.Is(err) {
+// 					// get the previous block from local
+// 					pb, _ = sc.GetBlock(ctx, b.PrevHash)
+// 				} else {
+// 					continue
+// 				}
 
-				if pb == nil {
-					continue
-				}
+// 				if pb == nil {
+// 					continue
+// 				}
 
-				// process previous block
-				if err := sc.processBlock(ctx, pb); err != nil {
-					logging.Logger.Error("process block, handle previous block failed",
-						zap.Int64("round", pb.Round),
-						zap.String("block", pb.Hash),
-						zap.Error(err))
-					continue
-				}
+// 				// process previous block
+// 				if err := sc.processBlock(ctx, pb); err != nil {
+// 					logging.Logger.Error("process block, handle previous block failed",
+// 						zap.Int64("round", pb.Round),
+// 						zap.String("block", pb.Hash),
+// 						zap.Error(err))
+// 					continue
+// 				}
 
-				// process this block again
-				if err := sc.processBlock(ctx, b); err != nil {
-					logging.Logger.Error("process block, failed after getting previous block",
-						zap.Int64("round", b.Round),
-						zap.String("block", b.Hash),
-						zap.Error(err))
-					continue
-				}
-			}
+// 				// process this block again
+// 				if err := sc.processBlock(ctx, b); err != nil {
+// 					logging.Logger.Error("process block, failed after getting previous block",
+// 						zap.Int64("round", b.Round),
+// 						zap.String("block", b.Hash),
+// 						zap.Error(err))
+// 					continue
+// 				}
+// 			}
 
-			lfbTk := sc.GetLatestLFBTicket(ctx)
-			lfb = sc.GetLatestFinalizedBlock()
-			logging.Logger.Debug("process block successfully",
-				zap.Int64("round", b.Round),
-				zap.Int64("lfb round", lfb.Round),
-				zap.Int64("lfb ticket round", lfbTk.Round))
+// 			lfbTk := sc.GetLatestLFBTicket(ctx)
+// 			lfb = sc.GetLatestFinalizedBlock()
+// 			logging.Logger.Debug("process block successfully",
+// 				zap.Int64("round", b.Round),
+// 				zap.Int64("lfb round", lfb.Round),
+// 				zap.Int64("lfb ticket round", lfbTk.Round))
 
-			if b.Round >= lfb.Round+aheadN || b.Round >= endRound {
-				syncing = false
-				if b.Round < lfbTk.Round {
-					logging.Logger.Debug("process block, hit end, trigger sync",
-						zap.Int64("round", b.Round),
-						zap.Int64("end round", endRound),
-						zap.Int64("current round", cr))
-					syncBlocksTimer.Reset(0)
-				}
-			}
-		}
-	}
-}
+// 			if b.Round >= lfb.Round+aheadN || b.Round >= endRound {
+// 				syncing = false
+// 				if b.Round < lfbTk.Round {
+// 					logging.Logger.Debug("process block, hit end, trigger sync",
+// 						zap.Int64("round", b.Round),
+// 						zap.Int64("end round", endRound),
+// 						zap.Int64("current round", cr))
+// 					syncBlocksTimer.Reset(0)
+// 				}
+// 			}
+// 		}
+// 	}
+// }
 
-func (sc *Chain) requestBlocks(ctx context.Context, startRound, reqNum int64) {
-	blocks := make([]*block.Block, reqNum)
-	wg := sync.WaitGroup{}
-	for i := int64(0); i < reqNum; i++ {
-		wg.Add(1)
-		go func(idx int64) {
-			defer wg.Done()
-			r := startRound + idx + 1
-			var cancel func()
-			cctx, cancel := context.WithTimeout(ctx, 8*time.Second)
-			defer cancel()
-			b, err := sc.GetNotarizedBlockFromSharders(cctx, "", r)
-			if err != nil {
-				// fetch from miners
-				b, err = sc.GetNotarizedBlock(cctx, "", r)
-				if err != nil {
-					logging.Logger.Error("request block failed",
-						zap.Int64("round", r),
-						zap.Error(err))
-					return
-				}
-			}
+// func (sc *Chain) requestBlocks(ctx context.Context, startRound, reqNum int64) {
+// 	blocks := make([]*block.Block, reqNum)
+// 	wg := sync.WaitGroup{}
+// 	for i := int64(0); i < reqNum; i++ {
+// 		wg.Add(1)
+// 		go func(idx int64) {
+// 			defer wg.Done()
+// 			r := startRound + idx + 1
+// 			var cancel func()
+// 			cctx, cancel := context.WithTimeout(ctx, 8*time.Second)
+// 			defer cancel()
+// 			b, err := sc.GetNotarizedBlockFromSharders(cctx, "", r)
+// 			if err != nil {
+// 				// fetch from miners
+// 				b, err = sc.GetNotarizedBlock(cctx, "", r)
+// 				if err != nil {
+// 					logging.Logger.Error("request block failed",
+// 						zap.Int64("round", r),
+// 						zap.Error(err))
+// 					return
+// 				}
+// 			}
 
-			blocks[idx] = b
-		}(i)
-	}
-	wg.Wait()
+// 			blocks[idx] = b
+// 		}(i)
+// 	}
+// 	wg.Wait()
 
-	for _, b := range blocks {
-		if b == nil {
-			// return if block is not acquired, break here as we will redo the sync process from the missed one later
-			return
-		}
+// 	for _, b := range blocks {
+// 		if b == nil {
+// 			// return if block is not acquired, break here as we will redo the sync process from the missed one later
+// 			return
+// 		}
 
-		if err := sc.PushToBlockProcessor(b); err != nil {
-			logging.Logger.Debug("requested block, but failed to pushed to process channel",
-				zap.Int64("round", b.Round), zap.Error(err))
-		}
-	}
-}
+// 		if err := sc.PushToBlockProcessor(b); err != nil {
+// 			logging.Logger.Debug("requested block, but failed to pushed to process channel",
+// 				zap.Int64("round", b.Round), zap.Error(err))
+// 		}
+// 	}
+// }
 
 func (sc *Chain) hasRoundSummary(ctx context.Context, rNum int64) (*round.Round, bool) {
 	r, err := sc.GetRoundFromStore(ctx, rNum)
